api/requirements.txt
requests==2.28.1
flask==2.1.2
flask-restx>=0.5.1
werkzeug==2.0.2
pyyaml==6.0
itsdangerous==2.1.2
SPARQLWrapper==2.0.0
ruamel.yaml==0.17.21
jinja2==3.1.2
black==22.12.0
coloredlogs==15.0.1
pytest~=7.2.0
pytest-cov==4.0.0
flask-cors==3.0.10
iteration_utilities==0.11.0